Development

+++ application/controllers/Assessments.php
@@ -29,7 +29,7 @@ class Assessments extends Easol_Controller {
                Version, AdministrationDate,
                edfi.AcademicSubjectType.CodeValue as Subject,
                edfi.GradeLevelType.CodeValue as Grade,
-			AVG(CAST(StudentAssessmentScoreResult.Result as INT)) as AverageResult,
+			AVG(CAST(StudentAssessmentScoreResult.Result as FLOAT)) as AverageResult,
 			COUNT(*) as StudentCount
 			FROM edfi.StudentAssessmentScoreResult
 			JOIN edfi.AcademicSubjectDescriptor ON edfi.AcademicSubjectDescriptor.AcademicSubjectDescriptorId = edfi.StudentAssessmentScoreResult
@@ -85,7 +85,7 @@ class Assessments extends Easol_Controller {
                Version, AdministrationDate,
                edfi.AcademicSubjectType.CodeValue as Subject,
                edfi.GradeLevelType.CodeValue as Grade,
-			AVG(CAST(StudentAssessmentScoreResult.Result as INT)) as AverageResult,
+			AVG(CAST(StudentAssessmentScoreResult.Result as FLOAT)) as AverageResult,
 			COUNT(*) as StudentCount
 			FROM edfi.StudentAssessmentScoreResult
 			JOIN edfi.AcademicSubjectDescriptor ON edfi.AcademicSubjectDescriptor.AcademicSubjectDescriptorId = edfi.StudentAssessmentScoreResult
